Se agrego AccesoDatos, se corrigio la query, se agrego el metodo eliminar fisico y se mejor el diseño del DGV

+++ TP2-Equipo-13A/dominio/Articulo.cs
@@ -15,7 +15,7 @@ namespace dominio
         public string Descripcion { get; set; }
         public Marca Marca { get; set; }
         public Categoria Categoria { get; set; }
-        public float Precio { get; set; }
+        public Decimal Precio { get; set; }
 
     }
 }
